pyproject.toml
[project]
name = "ViFoodKG"
version = "0.1.0"
description = "ViFoodKG: A Knowledge Graph for Vietnamese Cuisine Questions & Answers powered by Neo4j and Gemini"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "gugOfBoat" }
]
dependencies = [
    "wikipedia-api>=0.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "slugify>=0.0.1",
    "python-slugify>=8.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.6.0",
    "google-genai",
    "supabase>=2.0.0",
    "sentence-transformers>=2.2.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
